ai-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
scikit-learn==1.5.2
joblib==1.4.2
